division in translation units; cosmetic changes
